+++ src/main/java/com/sourav/majorProject/dao/FlightDao.java
@@ -4,6 +4,8 @@ import com.sourav.majorProject.model.Flight;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface FlightDao extends JpaRepository<Flight,Long> {
+public interface FlightDao extends JpaRepository<Flight, BigInteger> {
 }
