+++ src/main/java/com/sourav/majorProject/dao/ScheduleDao.java
@@ -4,6 +4,8 @@ import com.sourav.majorProject.model.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ScheduleDao extends JpaRepository<Schedule,Long> {
+public interface ScheduleDao extends JpaRepository<Schedule, BigInteger> {
 }
