+++ src/main/java/com/sourav/majorProject/dao/ScheduledFlightDao.java
@@ -4,6 +4,8 @@ import com.sourav.majorProject.model.ScheduledFlight;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ScheduledFlightDao extends JpaRepository<ScheduledFlight,Long> {
+public interface ScheduledFlightDao extends JpaRepository<ScheduledFlight, BigInteger> {
 }
